refactor: 💡 split format process into processors
