+++ src/processors/processor.ts
@@ -1,0 +1,8 @@
+import type Formatter from "src/formatter";
+
+export abstract class Processor {
+	constructor(protected formatter: Formatter) {}
+
+	abstract preProcess(content: string): Promise<any>;
+	abstract postProcess(content: string): Promise<any>;
+}
